@@ -12,44 +12,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/core": "^0.3.2",
-    "@angular/animations": "^5.2.11",
-    "@angular/common": "^5.2.11",
-    "@angular/compiler": "^5.2.11",
-    "@angular/core": "^5.2.11",
-    "@angular/forms": "^5.2.11",
-    "@angular/http": "^5.2.11",
-    "@angular/platform-browser": "^5.2.11",
-    "@angular/platform-browser-dynamic": "^5.2.11",
-    "@angular/router": "^5.2.11",
-    "@stomp/ng2-stompjs": "^4.0.0",
+    "@angular-devkit/core": "^0.6.8",
+    "@angular/animations": "^6.0.9",
+    "@angular/common": "^6.0.9",
+    "@angular/compiler": "^6.0.9",
+    "@angular/core": "^6.0.9",
+    "@angular/forms": "^6.0.9",
+    "@angular/http": "^6.0.9",
+    "@angular/platform-browser": "^6.0.9",
+    "@angular/platform-browser-dynamic": "^6.0.9",
+    "@angular/router": "^6.0.9",
+    "@stomp/ng2-stompjs": "^6.0.0",
     "bootstrap": "^3.3.7",
     "core-js": "^2.5.7",
     "jquery": "^3.3.1",
-    "moment": "^2.21.0",
-    "rxjs": "^5.5.11",
+    "moment": "^2.22.2",
+    "rxjs": "^6.2.2",
     "sockjs-client": "^1.1.5",
-    "webpack": "~3.11.0",
+    "webpack": "~4.16.1",
     "zone.js": "^0.8.20"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.3",
+    "@angular/cli": "^6.0.8",
     "@angular/compiler-cli": "^5.2.11",
     "@angular/language-service": "^5.2.11",
-    "@types/jasmine": "2.8.7",
-    "@types/node": "^10.1.4",
-    "codelyzer": "^4.2.1",
+    "@types/jasmine": "2.8.8",
+    "@types/node": "^10.5.2",
+    "codelyzer": "^4.4.2",
     "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^2.0.2",
+    "karma": "^2.0.4",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^2.0.0",
     "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^1.1.0",
+    "karma-jasmine-html-reporter": "^1.2.0",
     "protractor": "~5.3.2",
-    "ts-node": "~3.3.0",
-    "tslint": "~5.10.0",
-    "typescript": "~2.6.2"
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~2.6.0",
+    "@angular-devkit/build-angular": "~0.6.8"
   }
 }
