@@ -1,7 +1,7 @@
 /**
  * Created by shawn on 12/01/17.
  */
-import {BehaviorSubject, Observable} from "rxjs/Rx";
+import {BehaviorSubject, Observable} from "rxjs";
 import {Info} from "../models/info";
 import {Injectable} from "@angular/core";
 import {HttpClient} from "@angular/common/http";
