@@ -2,8 +2,7 @@
  * Created by shawn on 19/11/16.
  */
 import {Injectable} from "@angular/core";
-import {BehaviorSubject, Observable} from "rxjs/Rx";
-import "rxjs/add/operator/toPromise";
+import {BehaviorSubject, Observable} from "rxjs";
 import {Loco} from "../models/loco";
 import {HttpClient} from "@angular/common/http";
 
